Move UserPicker into common
UserPicker is a generic user-selection component used by both admin tools
(ChangeOwnerModal) and notifications (NotificationOwner). It only depends
on shared code (api, ui, common hooks), so it belongs in common rather
than inside the admin feature.

Fixes 2 module boundary violations (shared/notifications -> feature/admin).
